Set a location of a SIC model input

Set a location of a SIC model input

Usage

SicLocation(location)

Arguments

location

a list containing the location keys (see details)

Value

a SicLocation object which is a character string in the same format as the locations described in the sic documentation of PAR files: https://sic.g-eau.fr/Format-of-the-par-file.

See also

Examples

# How to impose 5 m3/s in the upstream offtake of the model?
# Define location of the boundary condition to set
locations <- SicLocations(list(Nd = 1, Pr = 1, car = "Q"))
# Define its value
sicInputs <- SicInput(5, locations = locations)

# How to impose an hydrograph at the upstream offtake of the model?
sicInputs <- SicInput(data.frame(t = c(0, 3600, 7200), # time in seconds
                                v = c(5, 20, 5)),     # flows
                      locations = locations)