Get correspondence between network object and columns in result binary file

Get correspondence between network object and columns in result binary file

Usage

get_result_tree(cfg, scenario, variant = 0)

Arguments

cfg

a config object. Configuration to use. See loadConfig for details

scenario

numeric, the scenario to use

variant

numeric, the variant to use (0 by default means no variant)

Value

a data.frame with following columns:

"bf", "sn", "nd", "pr", "ouv": location of the result with number of respectively reach, section, node, offtake, and device.

"var": the name of the calculated variable

"col": the column number in the matrix produced by read_bin_result_matrix

Warning

Up to now, this function only handle results at sections.

Examples

if (FALSE) { # \dontrun{
cfg <- cfg_tmp_project()
sic_run_steady(cfg, scenario = 1)
df <- get_result_tree(cfg, 1)
head(df)
} # }