Set project params in the SIC project XML file

Set project params in the SIC project XML file

Usage

set_project_params(params, cfg)

Arguments

params

list of parameters (See details)

cfg

a config object. Configuration to use. See loadConfig for details

Value

None. Use for side effect on the XML project file.

Details

The parameter params is a list of parameters to set in the SIC project XML file. The list should be named with the parameter name. Possible values of parameters are:

HomothetieX

HomothetieY

DX

ImportEdital

GeoUnit

GeoRefAngle