Run Talweg, Fluvia or Sirene for a configured project

Use sic_run_mesh to run the mesh generator, sic_run_steady for steady flow simulation, and sic_run_unsteady for unsteady flow simulation.

Usage

sic_run_mesh(cfg, params = list())

sic_run_steady(cfg, scenario, variant = 0, sicInputs = NULL, params = list())

sic_run_unsteady(
  cfg,
  scenario = iniParams[4],
  variant = iniParams[5],
  sicInputs = NULL,
  iniParams = NULL,
  params = list()
)

Arguments

cfg

a config object. Configuration to use. See loadConfig for details

params

list or character, see details

scenario

numeric, the scenario to use

variant

numeric, the variant to use (0 by default means no variant)

sicInputs

A SicInput object or a list of SicInput objects create by merge.SicInput used to create a PAR file injectig inputs for the simulation

iniParams

5-length numeric vector, see set_initial_conditions for details

Value

Error code returned by shell.

Details

The argument params handles the parameters describe in SIC documentation. If argument params is a list, arguments are injected in the command line by taking the items of the list with the conversion [key]=[value]. By default, the parameter INTERF=0 is added to the command line. If argument params is a character string it is directly used as parameters of the command line.

Examples

if (FALSE) {
# Set up the configuration model
cfg <- cfg_tmp_project()

# Generate the mesh of the model
sic_run_mesh(cfg)

# Run steady simulation for the scenario #1
sic_run_steady(cfg, scenario = 1)

# Import initial condition from scenario 1
# to scenario 1, variant 1 for unsteady flow simulation
set_initial_conditions(c(1, 0, 0, 1, 1), cfg = cfg)

# Run unsteady flow simulation
sic_run_unsteady(cfg, scenario = 1, variant = 1)

# Or initiate and run the same unsteady flow simulation in one call
sic_run_unsteady(cfg, iniParams = c(1, 0, 0, 1, 1))
}