Write a PAR file for SIC simulation

Write inputs in a PAR file respectively to https://sic.g-eau.fr/Format-of-the-par-file.

Usage

sic_write_par(cfg, scenario, sicInputs)

Arguments

cfg

a config object. Configuration to use. See loadConfig for details

scenario

numeric, the scenario to use

sicInputs

A SicInput object or a list of SicInput objects create by merge.SicInput used to create a PAR file injectig inputs for the simulation

Value

Nothing.

Examples

if (FALSE) {
# Setting the model configuration
cfg <- cfg_tmp_project()

# How to impose an hydrograph at the upstream offtake of the model?
# Define the location of the upstream boundary condition to set
locations <- SicLocations(list(Nd = 1, Pr = 1, car = "Q"))
# Define its time series
sicInputs <- SicInput(data.frame(t = c(0, 3600, 7200), # time in seconds
                                v = c(5, 20, 5)),     # flows
                      locations = locations)
# Write the parameters in the PAR file
sic_write_par(cfg, 1, sicInputs)
}