Tidy a result simulation

Tidy a result simulation

Usage

tidy_result(res)

compact_tidy_result(res)

Arguments

res

a SicResult matrix provided by get_result

Value

tidy_result returns a data.frame with one line by variable and by saved simulation result time step :

one column per object type of the result (example: "bf" and "sn" for a section or "nd" and "pr" for an offtake)

one column "var" for the definition of the result variable

one column "t" for the simulation time of the result variable

one columne "value" for the value of the result variable

compact_tidy_result returns a data.frame with one line by variable :

one column per object type of the result (example: "bf" and "sn" for a section or "nd" and "pr" for an offtake)

one column "var" for the definition of the result variable

one column "values" containing a data.frame with a column value

The data.frame contains an attribute "t" with the time of the saved simulation time steps in seconds.

Examples