Update abscissas of Strickler portions in a SIC model

As sic_import_reaches redefines abscissas in the reaches, and as the Stricklers are defined with abscissas, geometry importation can lead to an inconsistent model.

So the aim of this function is to reset Strickler definitions of the whole model to a default value with one Strickler portion by reach with correct abscissas.

Usage

update_portion_abscissas(cfg, stricklers = cfg$project$stricklers)

Arguments

cfg

a config object. Configuration to use. See loadConfig for details

stricklers

2-length numeric, Strickler coefficient to apply for minor bed and medium bed respectively

Value

Use for side effect on the XML project file.

Examples

if (FALSE) { # \dontrun{
cfg <- cfg_tmp_project()
update_portion_abscissas(cfg, KMin = 50)

# Display first Strickler definitions encountered in the XML file
x <- read_xml(cfg$project$path)
xml_find_first(x, "//Stricklers")
} # }